fix: @mocks/cookie.js -> ts
- jest에서 js파일 못읽어옴
